Variable 'caller_address' in the fault simulation indicator now is created unconditionally.
Now this variable is implemented without additional parameter of indicator.

kedr_stack_trace functionality moved from leak_check into KEDR and is used in the 'caller_address' variable implementation.

'calculator' functionality moved from fault simulation subdirectory into KEDR top-level directory.

All additional functionality, which implemented via source files and used for example modules, now use next principles:
-header files are placed in the include directory and marked as installed.
-source files do not configuration but include its own header files from KEDR include directory.
-example modules do not copy header files into each own directory but include them from KEDR include directory.

+++ sources/tools/capture_trace/kedr_capture_trace.c
@@ -1156,7 +1156,7 @@ filter_line_session(const char* str, size_t size,
     if(markers_counter == 0)
     {
         print_error("Trace line '%.*s' outside target session.",
-            str_end - str, str);
+            (int)(str_end - str), str);
         //only warning
         return 1;
     }
